% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdngenRsquare.R
\name{bdnmaxlik}
\alias{bdnmaxlik}
\title{Maximum likelihood estimation of linear regression using dbdn distribution}
\usage{
bdnmaxlik(
  data,
  t,
  startval,
  freeparid,
  extantonly = F,
  jittervar = 1e-04,
  optimmethod = "subplex",
  optimpar_plex = list(reltolpar = 1e-04, reltolval = 1e-05, abstolpar = 1e-07, maxiter =
    10000, num_cycles = 1),
  optimpar_DE = list(lower = -100, upper = 100, VTR = -Inf, strategy = 2, bs = FALSE, NP
    = 100, itermax = 200, CR = 0.5, F = 0.8, trace = T, initialpop = NULL, storepopfrom =
    1001, storepopfreq = 1, p = 0.2, c = 0, reltol = sqrt(.Machine$double.eps), steptol =
    200, parallelType = "none", cluster = NULL, packages = c(), parVar = c(), foreachArgs
    = list(), parallelArgs = NULL),
  optimpar_GenSA = list(lower = -100, upper = 100, maxit = 5000, nb.stop.improvement =
    1e+06, smooth = TRUE, max.call = 1e+07, max.time = 31540000, temperature = 5230,
    visiting.param = 2.62, acceptance.param = -5, simple.function = FALSE, trace.mat =
    TRUE, seed = -100377),
  convdetails = F
)
}
\arguments{
\item{data}{A data frame with columns n, tipspec and tipext, referring to number of species resulting from every tip after time t, estimated tip speciation and tip extinction rates}

\item{t}{Time period after which species are observed}

\item{startval}{Starting or fixed values for the optimisation of model parameters, lambda intercept, lambda slope, mu intercept and mu slope in this order, vector of length 4}

\item{freeparid}{Indices of model parameters startval that are free to be be optimised}

\item{extantonly}{Condition the distribution for non extinction, default is F}

\item{jittervar}{Standard deviation of a normal distribution by which lamb and mu are jittered in boundary situation}

\item{optimmethod}{Optimisation algorithm to use, default local "subplex" was sufficient in our testing, but "simplex"  or global "GenSA" or "DEoptim" can also be used}

\item{optimpar_plex}{List of optimisation parameters passed to local "subplex" or "simplex" routines if these are used}

\item{optimpar_DE}{List of optimisation parameters passed to "DE_optim" if this algorithm used}

\item{optimpar_GenSA}{List of optimisation parameters passed to "GenSA" if this algorithm used}

\item{convdetails}{Return convergence diagnostics, default is F}
}
\value{
A list of all parameter values par, free parameter values parexp, maximum log-likelihood llh and optionally convergence details
}
\description{
Estimates maximum likelihood values of linear regression parameters of a model where number of species after time t is drawn from a dbdn distribution with lambda and mu estimated as linear functions of tip observed lambda and mus.
}
\examples{
testdata=data.frame(n=1:10,tipspec=1:10,tipext=rep(0,10))

bdnmaxlikval=bdnmaxlik(data=testdata, t=1, startval=c(0,1,0,0),freeparid = 1:4)
bdnmaxlikval$par
bdnmaxlikval$llh

bdnsatloglikval=bdnsatloglik(testdata,1)
bdnsatloglikval

testdata2=data.frame(n=1:10,tipspec=1:10,tipext=20:11)

bdnmaxlikval=bdnmaxlik(data=testdata2, t=1, startval=c(0,1,0,0),freeparid = 1:4)
bdnmaxlikval$par
bdnmaxlikval$llh

bdnsatloglikval=bdnsatloglik(testdata,1)
bdnsatloglikval

restbdnmaxlikval=bdnmaxlik(data=testdata2, t=1, startval=c(0,0,0,0),freeparid = c(1,3,4))
restbdnmaxlikval
}
