% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdngenRsquare.R
\name{bdnregloglik}
\alias{bdnregloglik}
\title{Log-likelihood of linear transformation of tip-estimated lambda and mu using dbdn distribution}
\usage{
bdnregloglik(params, data, t, extantonly = F, jittervar = 1e-06)
}
\arguments{
\item{params}{A vector of length 4, with regression parameters ordered c(intercept_lambda, slope_lambda, intercept_mu, slope_mu)}

\item{data}{A data frame with columns n, tipspec and tipext, referring to number of species resulting from every tip after time t, estimated tip speciation and tip extinction rates}

\item{t}{Time period after which species are observed}

\item{extantonly}{Condition the distribution for non extinction, default is F}

\item{jittervar}{Standard deviation of a normal distribution by which lamb and mu are jittered in boundary situation}
}
\value{
Log-likelihood value
}
\description{
Calculates a log-likelihood as a sum of individual log probabilities for individual data points with real lambdas and mus being the linenar functions of observed tip lambdas and mus
}
\examples{
#dataset without extinct tips
testdata=data.frame(n=1:10,tipspec=1:10,tipext=rep(0,10))
bdnregloglik(c(0,1,0,0),testdata,1)
bdnregloglik(c(0,1,1,0),testdata,1)
bdnregloglik(c(0,1,0,1),testdata,1)
bdnregloglik(c(0,1,1,1),testdata,1)

#dataset with extinct tips
testdata0=data.frame(n=0:10,tipspec=0:10,tipext=rep(0,11))
bdnregloglik(c(0,1,1,0),testdata0,1, extantonly = T)
bdnregloglik(c(0,1,1,0),testdata0,1, extantonly = F)
bdnregloglik(c(0,1,1,0),testdata,1, extantonly = T)
bdnregloglik(c(0,1,1,0),testdata,1, extantonly = F)
}
