% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdngenRsquare.R
\name{bdnsatloglik}
\alias{bdnsatloglik}
\title{Saturated log-likelihood for the whole dataset}
\usage{
bdnsatloglik(
  data,
  mus = 0,
  t = 1,
  extantonly = F,
  jittervar = 1e-06,
  extinctmarginal = F,
  startval = 0.1,
  optimpar_plex = list(reltolpar = 1e-04, reltolval = 1e-05, abstolpar = 1e-07, maxiter =
    10000, num_cycles = 1)
)
}
\arguments{
\item{data}{A data frame with columns n, tipspec and tipext, referring to number of species resulting from every tip after time t, estimated tip speciation and tip extinction rates}

\item{t}{Time period after which species are observed}

\item{extantonly}{Condition the distribution for non extinction, default is F}

\item{jittervar}{Standard deviation of a normal distribution by which lamb and mu are jittered in boundary situation}

\item{extinctmarginal}{Do we marginalize the distribution over the assumed mu values, default is F}

\item{startval}{Starting values of the subplex optimisation of lambda in case we marginalize over assumed mu}

\item{optimpar_plex}{The control list of subplex optimisation of lambda in case we marginalize over assumed mu}

\item{mu}{Assumed extinction rates of a birth-death process in case we use extinction marginalisation}
}
\value{
Saturated log-likelihood value
}
\description{
Calculates a saturated log-likelihood as a sum of individual saturated log probabilities for individual data points, i.e. the maximum likelihood if lambda and mu are left free for every data point. Allows for marginalization of distributions using supplied mu values.
}
\examples{
testdata=data.frame(n=1:10,tipspec=1:10,tipext=rep(0,10))
bdnsatloglik(testdata,t=1,extinctmarginal = F)
bdnsatloglik(testdata,t=1,extinctmarginal = T)
bdnsatloglik(testdata,mus=1,t=1,extinctmarginal = T)
bdnsatloglik(testdata,mus=rep(1,10),t=1,extinctmarginal = T)
bdnsatloglik(testdata,t=1,extinctmarginal = T, startval = c(0,0,0,0,0,0,0,0,0,0))
bdnsatloglik(testdata,t=1,extinctmarginal = T, startval = c(-10,0,0,0,0,0,0,0,0,0))
bdnsatloglik(testdata,t=1,extinctmarginal = T, startval = c(0,0,0,0,0,0,0,0,0,-10))
}
