% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdngenRsquare.R
\name{bdntreesim}
\alias{bdntreesim}
\title{Simulation procedure for input data of bdn linear models}
\usage{
bdntreesim(lambdas, mus, ts, extantonly = F)
}
\arguments{
\item{lambdas}{Vector od speciation rates lambda}

\item{mus}{Vector of extinction rates mu}

\item{ts}{Vector of waiting times t}

\item{extantonly}{Iterate until a surviving tree is produced}
}
\value{
A data frame with columns n, tipspec, tipext and time, referring to number of species resulting from every tip after given  tip speciation, tip extinction rates, and waiting times
}
\description{
A function generating numbers data frame with numbers of species n resulting from a birth-death processes with given lambdas, mus and ts. All these three parameters are vectorized over.
}
\examples{
bdntreesim(rep(1,1000),rep(0,1000),rep(1,1000))
mean(bdntreesim(rep(1,1000),rep(0,1000),rep(1,1000))$n)
exp(1)

mean(bdntreesim(rep(2,1000),rep(0,1000),rep(1,1000))$n)
exp(2)

mean(bdntreesim(rep(2,1000),rep(1,1000),rep(1,1000))$n)
exp(1)
}
