% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdngenRsquare.R
\name{dbdn}
\alias{dbdn}
\title{Probability density function of number of species from a birth-death model}
\usage{
dbdn(n, lamb, mu, t, extantonly = F, log = F, jittervar = 1e-06)
}
\arguments{
\item{n}{Number of observed species}

\item{lamb}{Speciation rate of a birth-death process}

\item{mu}{Extinction rate of a birth-death process}

\item{t}{Time period after which species are observed}

\item{extantonly}{Condition the distribution for non extinction, default is F}

\item{log}{Return log-probability, default is F}

\item{jittervar}{Standard deviation of a normal distribution by which lamb and mu are jittered in boundary situation}
}
\value{
Probability of number of species for given alpha and mu
}
\description{
The function is taken from Luke Harmons book https://lukejharmon.github.io/pcm/pd/phylogeneticComparativeMethods.pdf, but with typo corrected according to Raup 1985 Paleobiology https://www.jstor.org/stable/2400422. All parameters are fully vectorized.
}
\examples{

#test for writing it well, with mu=1, t=1, the distribution should equal to geometric for n-1
dbdn(1,2,0,1)
dbdn(1,3,3,1)
dbdn(0,0,3,1)
dbdn(2,0,3,1)
dbdn(2,-1,3,1)
dbdn(2,0,3,5)
dbdn(2,5,5,5)
dbdn(1,2,0,1)
dgeom(1-1,exp(-2))

dbdn(10,2,0,1)
dgeom(10-1,exp(-2))

dbdn(1,12,0,1)
dgeom(1-1,exp(-12))

#tests for vectorisation
dbdn(1:5,1:5,0,1,log=c(F,T,T,T,T))
dbdn(2,2,0,1,F)

#negative values
dbdn(-2,12,0,1)
dbdn(1,-12,0,1)
dbdn(1,12,-10,1)
dbdn(0.1,12,-10,1)

#the distribution sums to 1
sum(dbdn(0:100000,1,2,1))
sum(dbdn(0:100000,1,0,1))
sum(dbdn(0:100000,1,0.5,10))

#extantonly testing
dbdn(1,1,2,1,extantonly = F)
dbdn(1,1,2,1, extantonly = T)
dbdn(0,1,2,1, extantonly = F)
dbdn(0,1,2,1, extantonly = T)
sum(dbdn(1:100000,1,2,1,extantonly = T))
sum(dbdn(1:100000,1,2,1,extantonly = F))
sum(dbdn(0,1,2,1,extantonly = F))
}
