% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdngenRsquare.R
\name{dbdnsatlinear}
\alias{dbdnsatlinear}
\title{Saturated log-likelihood for the whole dataset (with linear conditioning of extinction)}
\usage{
dbdnsatlinear(
  data,
  t,
  startvals,
  extantonly = F,
  jittervar = 1e-04,
  optimpar_plex = list(reltolpar = 1e-04, reltolval = 1e-05, abstolpar = 1e-07, maxiter =
    10000, num_cycles = 1),
  onlyfout = T
)
}
\arguments{
\item{data}{A data frame with columns n, tipspec and tipext, referring to number of species resulting from every tip after time t, estimated tip speciation and tip extinction rates}

\item{t}{Time period after which species are observed}

\item{startvals}{Starting values of the subplex optimisation of lambdas, mu intercept and mu slope in this order, vector of length n+2}

\item{extantonly}{Condition the distribution for non extinction, default is F}

\item{jittervar}{Standard deviation of a normal distribution by which lamb and mu are jittered in boundary situation}

\item{optimpar_plex}{The control list of subplex optimisation of lambdas and mu parameters}

\item{onlyfout}{Return just functional value or full output of optimisation, default is T}
}
\value{
Saturated log-likelihood value
}
\description{
Calculates a saturated log-likelihood as a sum of individual saturated log probabilities for individual data points, with marginalization of extinction as a maximum likelihood fitted linear function fo tip observed extinction values.
}
\examples{
testdata=data.frame(n=0:9,tipspec=1:10,tipext=rep(1,10))
bdnsatloglik(testdata,t=1,extinctmarginal = F)
bdnsatloglik(testdata,mus=testdata$tipext,t=1,extinctmarginal = T)
dbdnsatlinear(testdata, t=1, startvals=c(rep(0.1,length(testdata$n)),0.1,0.1), onlyfout = F)
}
