% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdngenRsquare.R
\name{optimglwrap}
\alias{optimglwrap}
\title{Wrapper for various local and global optimiser routines}
\usage{
optimglwrap(
  optimmethod = "simplex",
  optimpar_plex = list(reltolpar = 1e-04, reltolval = 1e-05, abstolpar = 1e-07, maxiter =
    10000, num_cycles = 1),
  optimpar_DE = list(lower = 0, upper = 1, VTR = -Inf, strategy = 2, bs = FALSE, NP =
    100, itermax = 200, CR = 0.5, F = 0.8, trace = T, initialpop = NULL, storepopfrom =
    1001, storepopfreq = 1, p = 0.2, c = 0, reltol = sqrt(.Machine$double.eps), steptol =
    200, parallelType = "none", cluster = NULL, packages = c(), parVar = c(), foreachArgs
    = list(), parallelArgs = NULL),
  optimpar_GenSA = list(lower = 0, upper = 1, maxit = 1000, nb.stop.improvement = 1e+06,
    smooth = TRUE, max.call = 1e+07, max.time = 31540000, temperature = 5230,
    visiting.param = 2.62, acceptance.param = -5, simple.function = FALSE, trace.mat =
    TRUE, seed = -100377),
  fc,
  trparsopt,
  ...
)
}
\arguments{
\item{optimmethod}{Optimisation algorithm to use, default is local "subplex", but "simplex"  or global "GenSA" or "DEoptim" can also be used}

\item{optimpar_plex}{List of optimisation parameters passed to local "subplex" or "simplex" routines if these are used, relative tolerance of parameters, relative tolerance of functional values, absolute tolerance of parameters, number of iterations and number of cycles for which the routine is run from the start can be set}

\item{optimpar_DE}{List of optimisation parameters passed to "DE_optim" if this algorithm used, see DEoptim package for details}

\item{optimpar_GenSA}{List of optimisation parameters passed to "GenSA" if this algorithm used, see GenSA package for details}

\item{fc}{Function to be optimised}

\item{trparsopt}{Starting values used in subplex, simplex or GenSA}
}
\value{
A list of generalized R square genRsquare,deviance of alternative model devalt, deviance of null model devnull, log-likelihood of alternative model loglikalt, log-likelihood of null model logliknul, saturated log-likelihood in the numerator of Rsquare formula logliksat1, saturated log-likelihood in the numerator of Rsquare formula logliksat2, regression parameters of alternative model paralt, regression parameters of null model parnul, and optionally convergence details
}
\description{
Wrapper function for subplex, simplex, differential evolution (DEoptim) and generalized simulated annealing (GenSA) optimisation routines.
}
